refactor: remove unused imports and simplify facet subquery parameters
